Lens.Micro.GHC.Internal

Description

Documentation

class IsByteString t where

Methods

packedBytes :: Lens' [Word8] t

Treat a list of bytes as a strict or lazy ByteString.

unpackedBytes :: Lens' t [Word8]

Treat a strict or lazy ByteString as a list of bytes.

packedChars :: Lens' String t

Treat a String as a strict or lazy ByteString. (Note that it will garble characters above 0xFF, same as pack does.)

unpackedChars :: Lens' t String

Treat a strict or lazy ByteString as a String. (Just as packedChars, it will garble characters above 0xFF.)

chars :: Traversal' t Char

Traverse characters in a strict or lazy ByteString (to traverse bytes instead of characters, use each).

Unpacking bytestrings

unpackStrict :: ByteString -> [Word8]

unpackStrict8 :: ByteString -> String

unpackLazy :: ByteString -> [Word8]

unpackLazy8 :: ByteString -> String

Converting bytestrings between strict and lazy

fromStrict :: ByteString -> ByteString

toStrict :: ByteString -> ByteString

Traversing bytestrings

traversedStrictTree :: Traversal' ByteString Word8

traversedStrictTree8 :: Traversal' ByteString Char

traversedLazy :: Traversal' ByteString Word8

traversedLazy8 :: Traversal' ByteString Char